test: add conftest.py
